move sql db files, update dependent imports and mocks
