+++ containers/ecr-viewer/src/app/services/listEcrDataService.ts
@@ -1,6 +1,6 @@
 import { formatDate, formatDateTime } from "@/app/services/formatDateService";
-import { get_pool } from "../api/services/sqlserver_db";
-import { getDB } from "../api/services/postgres_db";
+import { get_pool } from "../data/db/sqlserver_db";
+import { getDB } from "../data/db/postgres_db";
 import { DateRangePeriod } from "@/app/utils/date-utils";
 
 export interface CoreMetadataModel {
