+++ containers/ecr-viewer/src/app/tests/api/save-fhir-data-service.test.tsx
@@ -4,9 +4,9 @@
 
 import { saveMetadataToPostgres } from "../../api/save-fhir-data/save-fhir-data-service";
 import { BundleMetadata } from "../../api/save-fhir-data/types";
-import { getDB } from "../../api/services/postgres_db";
+import { getDB } from "../../data/db/postgres_db";
 
-jest.mock("../../api/services/postgres_db", () => ({
+jest.mock("../../data/db/postgres_db", () => ({
   getDB: jest.fn(),
 }));
